refactor: categorize images, delete unused images
