+++ src/images/skills-icons/figma-icon-mobile.png
